+++ src/account/operations/syncing/outputs.rs
@@ -70,53 +70,43 @@ impl AccountHandle {
 
     /// Gets outputs by their id, already known outputs are not requested again, but loaded from the account set as
     /// unspent, because we wouldn't get them from the node if they were spent
-    /// New requested output responses are returned and old ones separated with their accumulated balance
-    pub(crate) async fn get_outputs(
-        &self,
-        output_ids: Vec<OutputId>,
-        spent_outputs: bool,
-    ) -> crate::Result<(Vec<OutputResponse>, Vec<OutputResponse>)> {
+    pub(crate) async fn get_outputs(&self, output_ids: Vec<OutputId>) -> crate::Result<Vec<OutputResponse>> {
         log::debug!("[SYNC] start get_outputs");
         let get_outputs_start_time = Instant::now();
-        let mut found_outputs = Vec::new();
-        let mut loaded_outputs = Vec::new();
-        // For spent outputs we want to try to fetch all, so we can update them locally
-        if spent_outputs {
-            found_outputs = self.client.try_get_outputs(output_ids).await?;
-        } else {
-            let mut unknown_outputs = Vec::new();
-            let mut account = self.write().await;
-            let mut unspent_outputs = Vec::new();
-            for output_id in output_ids {
-                match account.outputs.get_mut(&output_id) {
-                    // set unspent
-                    Some(output_data) => {
-                        output_data.is_spent = false;
-                        unspent_outputs.push((output_id, output_data.clone()));
-                        loaded_outputs.push(OutputResponse {
-                            metadata: output_data.metadata.clone(),
-                            output: OutputDto::from(&output_data.output),
-                        });
-                    }
-                    None => unknown_outputs.push(output_id),
-                }
-            }
-            // known output is unspent, so insert it to the unspent outputs again, because if it was an
-            // alias/nft/foundry output it could have been removed when syncing without `sync_aliases_and_nfts`
-            for (output_id, output_data) in unspent_outputs {
-                account.unspent_outputs.insert(output_id, output_data);
-            }
+        let mut outputs = Vec::new();
 
-            if !unknown_outputs.is_empty() {
-                found_outputs = self.client.get_outputs(unknown_outputs).await?;
+        let mut unknown_outputs = Vec::new();
+        let mut account = self.write().await;
+        let mut unspent_outputs = Vec::new();
+        for output_id in output_ids {
+            match account.outputs.get_mut(&output_id) {
+                // set unspent
+                Some(output_data) => {
+                    output_data.is_spent = false;
+                    unspent_outputs.push((output_id, output_data.clone()));
+                    outputs.push(OutputResponse {
+                        metadata: output_data.metadata.clone(),
+                        output: OutputDto::from(&output_data.output),
+                    });
+                }
+                None => unknown_outputs.push(output_id),
             }
+        }
+        // known output is unspent, so insert it to the unspent outputs again, because if it was an
+        // alias/nft/foundry output it could have been removed when syncing without `sync_aliases_and_nfts`
+        for (output_id, output_data) in unspent_outputs {
+            account.unspent_outputs.insert(output_id, output_data);
+        }
+
+        if !unknown_outputs.is_empty() {
+            outputs.extend(self.client.get_outputs(unknown_outputs).await?);
         }
 
         log::debug!(
             "[SYNC] finished get_outputs in {:.2?}",
             get_outputs_start_time.elapsed()
         );
-        Ok((found_outputs, loaded_outputs))
+        Ok(outputs)
     }
 
     // Try to get transactions and inputs for received outputs